feat: Implement backend structure with database connection, authentication, and CRUD operations
- Added database connection setup using MySQL with connection pooling.
- Created authentication middleware for token verification and role-based access control.
- Developed routes for managing appointments, users (registration and login), patients, practitioners, slots, staff, treatment plans, and file uploads.
- Implemented health check route for server status.
- Added JSON data for therapies with required stock items.
- Integrated error handling and logging for better debugging and user feedback.
