+++ backend/server.js
@@ -3,12 +3,18 @@ const cors = require("cors");
 const helmet = require("helmet");
 const rateLimit = require("express-rate-limit");
 const dotenv = require("dotenv");
-const mysql = require("mysql2/promise");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const { v4: uuidv4 } = require("uuid");
-const multer = require("multer");
-const nodemailer = require("nodemailer");
+const { initializeDatabase } = require("./db");
+
+// Import route modules
+const authRoutes = require("./routes/auth");
+const patientRoutes = require("./routes/patients");
+const practitionerRoutes = require("./routes/practitioners");
+const appointmentRoutes = require("./routes/appointments");
+const treatmentPlanRoutes = require("./routes/treatmentPlans");
+const staffRoutes = require("./routes/staff");
+const uploadRoutes = require("./routes/uploads");
+const slotRoutes = require("./routes/slots");
+const healthRoutes = require("./routes/health");
 
 // Load environment variables
 dotenv.config();
@@ -16,31 +22,6 @@ dotenv.config();
 // Initialize Express app
 const app = express();
 const PORT = process.env.PORT || 5000;
-
-// Database connection
-const dbConfig = {
-  host: process.env.DB_HOST || "localhost",
-  user: process.env.DB_USER || "root",
-  password: process.env.DB_PASSWORD || "",
-  database: process.env.DB_NAME || "panchakarma_db",
-  port: process.env.DB_PORT || 3306,
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 0,
-};
-
-let pool;
-
-// Initialize database connection
-async function initializeDatabase() {
-  try {
-    pool = mysql.createPool(dbConfig);
-    console.log("Database connected successfully");
-  } catch (error) {
-    console.error("Database connection failed:", error);
-    process.exit(1);
-  }
-}
 
 // Middleware
 app.use(helmet());
@@ -55,1000 +36,16 @@ const limiter = rateLimit({
 });
 app.use(limiter);
 
-// File upload configuration
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, "uploads/");
-  },
-  filename: (req, file, cb) => {
-    cb(null, Date.now() + "-" + file.originalname);
-  },
-});
-const upload = multer({ storage });
-
-// Authentication middleware
-const authenticateToken = (req, res, next) => {
-  const authHeader = req.headers["authorization"];
-  const token = authHeader && authHeader.split(" ")[1];
-
-  if (!token) return res.status(401).json({ error: "Access token required" });
-
-  jwt.verify(
-    token,
-    process.env.JWT_SECRET || "your-secret-key",
-    (err, user) => {
-      if (err) return res.status(403).json({ error: "Invalid token" });
-      req.user = user;
-      next();
-    }
-  );
-};
-
-// Role-based authorization middleware
-const authorizeRoles = (...roles) => {
-  return (req, res, next) => {
-    if (!roles.includes(req.user.role)) {
-      return res.status(403).json({ error: "Insufficient permissions" });
-    }
-    next();
-  };
-};
-
 // Routes
-
-// User Authentication Routes
-app.post("/api/auth/register", async (req, res) => {
-  try {
-    const { email, phone, password, firstName, lastName, userType } = req.body;
-    console.log("Registering user:", email, phone, userType);
-
-    // Validate user type
-    const validUserTypes = ["patient", "practitioner", "admin", "staff"];
-    if (!validUserTypes.includes(userType)) {
-      return res.status(400).json({ error: "Invalid user type" });
-    }
-
-    // Check if user already exists
-    const [existingUser] = await pool.execute(
-      "SELECT * FROM users WHERE email = ? OR phone = ?",
-      [email, phone]
-    );
-
-    if (existingUser.length > 0) {
-      return res.status(400).json({ error: "User already exists" });
-    }
-
-    // Hash password
-    const saltRounds = 12;
-    const hashedPassword = await bcrypt.hash(password, saltRounds);
-
-    // Create user
-    const userId = uuidv4();
-    const [newUser] = await pool.execute(
-      "INSERT INTO users (user_id, email, phone, password_hash, user_type, first_name, last_name) VALUES (?, ?, ?, ?, ?, ?, ?)",
-      [userId, email, phone, hashedPassword, userType, firstName, lastName]
-    );
-
-    res.status(201).json({
-      message: "User registered successfully",
-      user: {
-        userId: userId,
-        email: email,
-        userType: userType,
-      },
-    });
-  } catch (error) {
-    console.error("Registration error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-app.post("/api/auth/login", async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    // Find user
-    const [user] = await pool.execute("SELECT * FROM users WHERE email = ?", [
-      email,
-    ]);
-    if (user.length === 0) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-
-    // Check password
-    const validPassword = await bcrypt.compare(password, user[0].password_hash);
-    if (!validPassword) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-
-    // Generate JWT token
-    const token = jwt.sign(
-      {
-        userId: user[0].user_id,
-        email: user[0].email,
-        role: user[0].user_type,
-      },
-      process.env.JWT_SECRET || "your-secret-key",
-      { expiresIn: "24h" }
-    );
-
-    res.json({
-      message: "Login successful",
-      token,
-      user: {
-        userId: user[0].user_id,
-        email: user[0].email,
-        firstName: user[0].first_name,
-        lastName: user[0].last_name,
-        userType: user[0].user_type,
-      },
-    });
-  } catch (error) {
-    console.error("Login error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Patient Management Routes
-app.get(
-  "/api/patients",
-  authenticateToken,
-  async (req, res) => {
-    try {
-      if (req.user.role === 'patient') {
-        // Patients can only see their own profile
-        const [patients] = await pool.execute(`
-          SELECT p.*, u.first_name, u.last_name, u.email, u.phone
-          FROM patients p
-          JOIN users u ON p.user_id = u.user_id
-          WHERE p.user_id = ?
-        `, [req.user.userId]);
-        res.json(patients);
-      } else {
-        // Admin and practitioners can see all patients
-        const [patients] = await pool.execute(`
-          SELECT p.*, u.first_name, u.last_name, u.email, u.phone
-          FROM patients p
-          JOIN users u ON p.user_id = u.user_id
-          ORDER BY p.created_at DESC
-        `);
-        res.json(patients);
-      }
-    } catch (error) {
-      console.error("Get patients error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-app.post("/api/patients", authenticateToken, async (req, res) => {
-  try {
-    const {
-      dateOfBirth,
-      gender,
-      bloodGroup,
-      height,
-      weight,
-      occupation,
-      maritalStatus,
-      emergencyContactName,
-      emergencyContactPhone,
-      emergencyContactRelationship,
-      medicalConditions,
-      allergies,
-      currentMedications,
-      pastSurgeries,
-      familyMedicalHistory,
-      lifestyleHabits,
-      dietaryPreferences,
-      exerciseRoutine,
-      prakritiAssessment,
-      vikritiAssessment,
-      doshaDominance,
-    } = req.body;
-
-    const patientId = uuidv4();
-    await pool.execute(
-      `INSERT INTO patients (
-        patient_id, user_id, date_of_birth, gender, blood_group, height_cm, weight_kg,
-        occupation, marital_status, emergency_contact_name, emergency_contact_phone,
-        emergency_contact_relationship, medical_conditions, allergies, current_medications,
-        past_surgeries, family_medical_history, lifestyle_habits, dietary_preferences,
-        exercise_routine, prakriti_assessment, vikriti_assessment, dosha_dominance
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-      [
-        patientId,
-        req.user.userId,
-        dateOfBirth,
-        gender,
-        bloodGroup,
-        height,
-        weight,
-        occupation,
-        maritalStatus,
-        emergencyContactName,
-        emergencyContactPhone,
-        emergencyContactRelationship,
-        medicalConditions,
-        allergies,
-        currentMedications,
-        pastSurgeries,
-        familyMedicalHistory,
-        lifestyleHabits,
-        dietaryPreferences,
-        exerciseRoutine,
-        JSON.stringify(prakritiAssessment),
-        JSON.stringify(vikritiAssessment),
-        doshaDominance,
-      ]
-    );
-
-    const [newPatient] = await pool.execute(
-      `SELECT * FROM patients WHERE patient_id = ?`,
-      [patientId]
-    );
-
-    res.status(201).json({
-      message: "Patient profile created successfully",
-      patient: newPatient[0],
-    });
-  } catch (error) {
-    console.error("Create patient error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Practitioner Management Routes
-app.get("/api/practitioners", authenticateToken, async (req, res) => {
-  try {
-    const [practitioners] = await pool.execute(`
-      SELECT p.*, u.first_name, u.last_name, u.email, u.phone
-      FROM practitioners p
-      JOIN users u ON p.user_id = u.user_id
-      WHERE p.verification_status = 'verified'
-      ORDER BY p.created_at DESC
-    `);
-    console.log('Practitioners query result:', practitioners);
-    res.json(practitioners);
-  } catch (error) {
-    console.error("Get practitioners error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-app.post(
-  "/api/practitioners",
-  authenticateToken,
-  authorizeRoles("admin"),
-  async (req, res) => {
-    try {
-      const {
-        licenseNumber,
-        qualification,
-        specializations,
-        experienceYears,
-        languagesSpoken,
-        consultationFee,
-        clinicAffiliation,
-        practiceStartDate,
-        workingHours,
-        consultationDuration,
-        maxPatientsPerDay,
-      } = req.body;
-
-      const practitionerId = uuidv4();
-      await pool.execute(
-        `INSERT INTO practitioners (
-        practitioner_id, user_id, license_number, qualification, specializations,
-        experience_years, languages_spoken, consultation_fee, clinic_affiliation,
-        practice_start_date, working_hours, consultation_duration, max_patients_per_day
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-        [
-          practitionerId,
-          req.user.userId,
-          licenseNumber,
-          qualification,
-          JSON.stringify(specializations),
-          experienceYears,
-          JSON.stringify(languagesSpoken),
-          consultationFee,
-          clinicAffiliation,
-          practiceStartDate,
-          JSON.stringify(workingHours),
-          consultationDuration,
-          maxPatientsPerDay,
-        ]
-      );
-
-      const [newPractitioner] = await pool.execute(
-        `SELECT * FROM practitioners WHERE practitioner_id = ?`,
-        [practitionerId]
-      );
-
-      res.status(201).json({
-        message: "Practitioner profile created successfully",
-        practitioner: newPractitioner[0],
-      });
-    } catch (error) {
-      console.error("Create practitioner error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Get current practitioner's profile
-app.get(
-  "/api/practitioners/profile",
-  authenticateToken,
-  authorizeRoles("practitioner"),
-  async (req, res) => {
-    try {
-      const [rows] = await pool.execute(
-        `SELECT p.*, u.first_name, u.last_name, u.email, u.phone
-         FROM practitioners p
-         JOIN users u ON p.user_id = u.user_id
-         WHERE p.user_id = ?`,
-        [req.user.userId]
-      );
-
-      if (rows.length === 0) {
-        return res
-          .status(404)
-          .json({ error: "Practitioner profile not found" });
-      }
-
-      res.json(rows[0]);
-    } catch (error) {
-      console.error("Get practitioner profile error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Update current practitioner's profile
-app.put(
-  "/api/practitioners/profile",
-  authenticateToken,
-  authorizeRoles("practitioner"),
-  async (req, res) => {
-    try {
-      const { workingHours } = req.body;
-
-      await pool.execute(
-        `
-      UPDATE practitioners
-      SET working_hours = ?, updated_at = CURRENT_TIMESTAMP
-      WHERE user_id = ?
-    `,
-        [JSON.stringify(workingHours), req.user.userId]
-      );
-
-      const [updatedPractitioner] = await pool.execute(
-        `SELECT * FROM practitioners WHERE user_id = ?`,
-        [req.user.userId]
-      );
-
-      if (updatedPractitioner.length === 0) {
-        return res
-          .status(404)
-          .json({ error: "Practitioner profile not found" });
-      }
-
-      res.json({
-        message: "Profile updated successfully",
-        practitioner: updatedPractitioner[0],
-      });
-    } catch (error) {
-      console.error("Update practitioner profile error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Appointment Management Routes
-app.get("/api/appointments", authenticateToken, async (req, res) => {
-  try {
-    let query;
-    let params = [];
-
-    if (req.user.role === "patient") {
-      // Get patient's appointments
-      query = `
-        SELECT a.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM appointments a
-        JOIN patients pt ON a.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON a.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        WHERE pt.user_id = ?
-        ORDER BY a.appointment_date DESC, a.start_time DESC
-      `;
-      params = [req.user.userId];
-    } else if (req.user.role === "practitioner") {
-      // Get practitioner's appointments
-      query = `
-        SELECT a.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM appointments a
-        JOIN patients pt ON a.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON a.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        WHERE prac.user_id = ?
-        ORDER BY a.appointment_date DESC, a.start_time DESC
-      `;
-      params = [req.user.userId];
-    } else {
-      // Admin gets all appointments
-      query = `
-        SELECT a.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM appointments a
-        JOIN patients pt ON a.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON a.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        ORDER BY a.appointment_date DESC, a.start_time DESC
-      `;
-    }
-
-    const [appointments] = await pool.execute(query, params);
-    res.json(appointments);
-  } catch (error) {
-    console.error("Get appointments error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-app.post("/api/appointments", authenticateToken, async (req, res) => {
-  try {
-    const {
-      practitionerId,
-      appointmentDate,
-      startTime,
-      endTime,
-      serviceType,
-      consultationType,
-      specialInstructions,
-      preparationNotes,
-    } = req.body;
-
-    // Get patient ID from user ID
-    const [patient] = await pool.execute(
-      "SELECT patient_id FROM patients WHERE user_id = ?",
-      [req.user.userId]
-    );
-    if (patient.length === 0) {
-      return res.status(400).json({ error: "Patient profile not found" });
-    }
-
-    const appointmentId = uuidv4();
-
-    await pool.execute(
-      `INSERT INTO appointments (
-        appointment_id, patient_id, practitioner_id, appointment_date, start_time, end_time,
-        service_type, consultation_type, special_instructions, preparation_notes,
-        booking_channel, confirmation_code
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-      [
-        appointmentId,
-        patient[0].patient_id,
-        practitionerId,
-        appointmentDate,
-        startTime,
-        endTime,
-        serviceType,
-        consultationType,
-        specialInstructions,
-        preparationNotes,
-        'app',
-        Math.random().toString(36).substring(2, 10).toUpperCase(),
-      ]
-    );
-
-    const [newAppointment] = await pool.execute(
-      `SELECT * FROM appointments WHERE appointment_id = ?`,
-      [appointmentId]
-    );
-
-    res.status(201).json({
-      message: "Appointment booked successfully",
-      appointment: newAppointment[0],
-    });
-  } catch (error) {
-    console.error("Create appointment error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Treatment Plan Routes
-app.get("/api/treatment-plans", authenticateToken, async (req, res) => {
-  try {
-    let query;
-    let params = [];
-
-    if (req.user.role === "patient") {
-      query = `
-        SELECT tp.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM treatment_plans tp
-        JOIN patients pt ON tp.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON tp.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        WHERE pt.user_id = ?
-        ORDER BY tp.created_at DESC
-      `;
-      params = [req.user.userId];
-    } else if (req.user.role === "practitioner") {
-      query = `
-        SELECT tp.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM treatment_plans tp
-        JOIN patients pt ON tp.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON tp.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        WHERE prac.user_id = ?
-        ORDER BY tp.created_at DESC
-      `;
-      params = [req.user.userId];
-    } else {
-      query = `
-        SELECT tp.*, p.first_name as patient_first_name, p.last_name as patient_last_name,
-               pr.first_name as practitioner_first_name, pr.last_name as practitioner_last_name
-        FROM treatment_plans tp
-        JOIN patients pt ON tp.patient_id = pt.patient_id
-        JOIN users p ON pt.user_id = p.user_id
-        JOIN practitioners prac ON tp.practitioner_id = prac.practitioner_id
-        JOIN users pr ON prac.user_id = pr.user_id
-        ORDER BY tp.created_at DESC
-      `;
-    }
-
-    const [treatmentPlans] = await pool.execute(query, params);
-    res.json(treatmentPlans);
-  } catch (error) {
-    console.error("Get treatment plans error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-app.post(
-  "/api/treatment-plans",
-  authenticateToken,
-  authorizeRoles("practitioner", "admin"),
-  async (req, res) => {
-    try {
-      const {
-        patientId,
-        treatmentName,
-        treatmentType,
-        startDate,
-        endDate,
-        totalSessions,
-        purvakarmaProtocols,
-        pradhanakarmaProtocols,
-        paschatkarmaProtocols,
-        contraindications,
-        expectedOutcomes,
-        specialInstructions,
-        totalCost,
-      } = req.body;
-
-      const treatmentPlanId = uuidv4();
-      await pool.execute(
-        `INSERT INTO treatment_plans (
-        treatment_plan_id, patient_id, practitioner_id, treatment_name, treatment_type,
-        start_date, end_date, total_sessions, purvakarma_protocols, pradhanakarma_protocols,
-        paschatkarma_protocols, contraindications, expected_outcomes, special_instructions,
-        total_cost
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-        [
-          treatmentPlanId,
-          patientId,
-          req.user.userId,
-          treatmentName,
-          treatmentType,
-          startDate,
-          endDate,
-          totalSessions,
-          JSON.stringify(purvakarmaProtocols),
-          JSON.stringify(pradhanakarmaProtocols),
-          JSON.stringify(paschatkarmaProtocols),
-          contraindications,
-          JSON.stringify(expectedOutcomes),
-          specialInstructions,
-          totalCost,
-        ]
-      );
-
-      const [newTreatmentPlan] = await pool.execute(
-        `SELECT * FROM treatment_plans WHERE treatment_plan_id = ?`,
-        [treatmentPlanId]
-      );
-
-      res.status(201).json({
-        message: "Treatment plan created successfully",
-        treatmentPlan: newTreatmentPlan[0],
-      });
-    } catch (error) {
-      console.error("Create treatment plan error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Staff Routes
-
-// Get all stock items
-app.get(
-  "/api/staff/stock",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const [stock] = await pool.execute(`
-      SELECT s.*, u.first_name, u.last_name 
-      FROM stock s 
-      JOIN users u ON s.updated_by = u.user_id 
-      ORDER BY s.last_updated DESC
-    `);
-      res.json(stock);
-    } catch (error) {
-      console.error("Get stock error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Add new stock item
-app.post(
-  "/api/staff/stock",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const { itemName, quantity, unit } = req.body;
-      const stockId = uuidv4();
-
-      await pool.execute(
-        "INSERT INTO stock (id, item_name, quantity, unit, updated_by) VALUES (?, ?, ?, ?, ?)",
-        [stockId, itemName, quantity, unit, req.user.userId]
-      );
-
-      const [newStock] = await pool.execute(
-        `
-      SELECT s.*, u.first_name, u.last_name 
-      FROM stock s 
-      JOIN users u ON s.updated_by = u.user_id 
-      WHERE s.id = ?
-    `,
-        [stockId]
-      );
-
-      res.status(201).json({
-        message: "Stock item added successfully",
-        stock: newStock[0],
-      });
-    } catch (error) {
-      console.error("Add stock error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Update stock item
-app.put(
-  "/api/staff/stock/:id",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const { id } = req.params;
-      const { itemName, quantity, unit } = req.body;
-
-      await pool.execute(
-        "UPDATE stock SET item_name = ?, quantity = ?, unit = ?, updated_by = ? WHERE id = ?",
-        [itemName, quantity, unit, req.user.userId, id]
-      );
-
-      const [updatedStock] = await pool.execute(
-        `
-      SELECT s.*, u.first_name, u.last_name 
-      FROM stock s 
-      JOIN users u ON s.updated_by = u.user_id 
-      WHERE s.id = ?
-    `,
-        [id]
-      );
-
-      if (updatedStock.length === 0) {
-        return res.status(404).json({ error: "Stock item not found" });
-      }
-
-      res.json({
-        message: "Stock item updated successfully",
-        stock: updatedStock[0],
-      });
-    } catch (error) {
-      console.error("Update stock error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Delete stock item
-app.delete(
-  "/api/staff/stock/:id",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const { id } = req.params;
-
-      // Check if stock item exists
-      const [existingStock] = await pool.execute(
-        "SELECT * FROM stock WHERE id = ?",
-        [id]
-      );
-      if (existingStock.length === 0) {
-        return res.status(404).json({ error: "Stock item not found" });
-      }
-
-      // Delete the stock item
-      await pool.execute("DELETE FROM stock WHERE id = ?", [id]);
-
-      res.json({
-        message: "Stock item deleted successfully",
-      });
-    } catch (error) {
-      console.error("Delete stock error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Get all rooms
-app.get(
-  "/api/staff/rooms",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const [rooms] = await pool.execute(`
-      SELECT r.*, u.first_name, u.last_name 
-      FROM rooms r 
-      JOIN users u ON r.last_updated_by = u.user_id 
-      ORDER BY r.room_name
-    `);
-      res.json(rooms);
-    } catch (error) {
-      console.error("Get rooms error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// Update room status
-app.put(
-  "/api/staff/rooms/:id",
-  authenticateToken,
-  authorizeRoles("staff", "admin"),
-  async (req, res) => {
-    try {
-      const { id } = req.params;
-      const { status } = req.body;
-
-      const validStatuses = ["available", "occupied", "maintenance"];
-      if (!validStatuses.includes(status)) {
-        return res.status(400).json({ error: "Invalid status" });
-      }
-
-      await pool.execute(
-        "UPDATE rooms SET status = ?, last_updated_by = ? WHERE id = ?",
-        [status, req.user.userId, id]
-      );
-
-      const [updatedRoom] = await pool.execute(
-        `
-      SELECT r.*, u.first_name, u.last_name 
-      FROM rooms r 
-      JOIN users u ON r.last_updated_by = u.user_id 
-      WHERE r.id = ?
-    `,
-        [id]
-      );
-
-      if (updatedRoom.length === 0) {
-        return res.status(404).json({ error: "Room not found" });
-      }
-
-      res.json({
-        message: "Room status updated successfully",
-        room: updatedRoom[0],
-      });
-    } catch (error) {
-      console.error("Update room error:", error);
-      res.status(500).json({ error: "Internal server error" });
-    }
-  }
-);
-
-// File upload route
-app.post(
-  "/api/upload",
-  authenticateToken,
-  upload.single("file"),
-  (req, res) => {
-    if (!req.file) {
-      return res.status(400).json({ error: "No file uploaded" });
-    }
-    res.json({
-      message: "File uploaded successfully",
-      filePath: req.file.path,
-      filename: req.file.filename,
-    });
-  }
-);
-
-// Slots Management Routes
-
-// Generate slots for a practitioner
-app.post("/api/slots/generate/:practitionerId", authenticateToken, authorizeRoles("admin", "practitioner"), async (req, res) => {
-  try {
-    const { practitionerId } = req.params;
-    const { regenerate } = req.body;
-
-    // Import the slot generator
-    const { generatePractitionerSlots } = require('./slotGenerator');
-
-    await generatePractitionerSlots(practitionerId, regenerate || false, pool);
-
-    res.json({ message: "Slots generated successfully" });
-  } catch (error) {
-    console.error("Generate slots error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Get slots for a practitioner by day and status
-app.get("/api/slots/:practitionerId", authenticateToken, async (req, res) => {
-  try {
-    const { practitionerId } = req.params;
-    const { day, status } = req.query;
-
-    let query = "SELECT * FROM slots WHERE practitioner_id = ?";
-    let params = [practitionerId];
-
-    if (day) {
-      query += " AND day = ?";
-      params.push(day.toLowerCase());
-    }
-
-    if (status) {
-      query += " AND status = ?";
-      params.push(status);
-    }
-
-    query += " ORDER BY day, start_time";
-
-    let [slots] = await pool.execute(query, params);
-
-    // If no slots found for the specific day, try to auto-generate them
-    if (slots.length === 0 && day) {
-      try {
-        const { generatePractitionerSlots } = require('./slotGenerator');
-        await generatePractitionerSlots(practitionerId, false, pool);
-
-        // Re-query after generation
-        [slots] = await pool.execute(query, params);
-      } catch (genError) {
-        console.error('Error auto-generating slots:', genError);
-        // Continue with empty slots if generation fails
-      }
-    }
-
-    res.json(slots);
-  } catch (error) {
-    console.error("Get slots error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Update slot status
-app.put("/api/slots/:slotId", authenticateToken, async (req, res) => {
-  try {
-    const { slotId } = req.params;
-    const { status } = req.body;
-
-    const validStatuses = ["booked", "free", "leave"];
-    if (!validStatuses.includes(status)) {
-      return res.status(400).json({ error: "Invalid status" });
-    }
-
-    const [result] = await pool.execute(
-      "UPDATE slots SET status = ?, updated_at = CURRENT_TIMESTAMP WHERE slot_id = ?",
-      [status, slotId]
-    );
-
-    if (result.affectedRows === 0) {
-      return res.status(404).json({ error: "Slot not found" });
-    }
-
-    const [updatedSlot] = await pool.execute("SELECT * FROM slots WHERE slot_id = ?", [slotId]);
-
-    res.json({
-      message: "Slot updated successfully",
-      slot: updatedSlot[0]
-    });
-  } catch (error) {
-    console.error("Update slot error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Get available slots across all practitioners for a specific date and time
-app.post("/api/slots/available", authenticateToken, async (req, res) => {
-  try {
-    const { date, time } = req.body;
-
-    if (!date || !time) {
-      return res.status(400).json({ error: "Date and time are required" });
-    }
-
-    // Get day of week from date
-    const dateObj = new Date(date);
-    const days = ['sunday', 'monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday'];
-    const dayOfWeek = days[dateObj.getDay()];
-
-    console.log('Searching for available doctors:', { date, time, dayOfWeek });
-
-    // Get all practitioners
-    const [practitioners] = await pool.execute(`
-      SELECT 
-        p.practitioner_id,
-        first_name,
-        last_name,
-        p.specializations
-      FROM practitioners p
-      natural join users
-    `);
-
-    console.log('Total practitioners:', practitioners.length);
-
-    // For each practitioner, check if they have a booked slot at the specified time
-    const availablePractitioners = [];
-
-    for (const practitioner of practitioners) {
-      // Check if this practitioner has a booked appointment at the specified date and time
-      const [bookedAppointments] = await pool.execute(`
-        SELECT COUNT(*) as booked_count
-        FROM appointments
-        WHERE practitioner_id = ?
-          AND appointment_date = ?
-          AND start_time LIKE ?
-          AND status IN ('scheduled', 'confirmed')
-      `, [practitioner.practitioner_id, date, `${time}%`]);
-      console.log(bookedAppointments);
-      if (bookedAppointments[0].booked_count === 0) {
-        availablePractitioners.push(practitioner);
-      }
-    }
-
-    console.log('Available practitioners:', availablePractitioners.length);
-    res.json(availablePractitioners);
-  } catch (error) {
-    console.error("Get available doctors error:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
-
-// Health check route
-app.get("/api/health", (req, res) => {
-  res.json({ status: "OK", timestamp: new Date().toISOString() });
-});
+app.use("/api/auth", authRoutes);
+app.use("/api/patients", patientRoutes);
+app.use("/api/practitioners", practitionerRoutes);
+app.use("/api/appointments", appointmentRoutes);
+app.use("/api/treatment-plans", treatmentPlanRoutes);
+app.use("/api/staff", staffRoutes);
+app.use("/api/upload", uploadRoutes);
+app.use("/api/slots", slotRoutes);
+app.use("/api/health", healthRoutes);
 
 // Error handling middleware
 app.use((err, req, res, next) => {
