+++ frontend/public/data/therapies.json
@@ -1,0 +1,182 @@
+[
+    {
+        "id": 1,
+        "name": "Abhyanga (Oil Massage)",
+        "requiredItems": [
+            {
+                "stockId": 1,
+                "quantity": 200
+            },
+            {
+                "stockId": 16,
+                "quantity": 2
+            },
+            {
+                "stockId": 12,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 2,
+        "name": "Shirodhara",
+        "requiredItems": [
+            {
+                "stockId": 1,
+                "quantity": 500
+            },
+            {
+                "stockId": 13,
+                "quantity": 1
+            },
+            {
+                "stockId": 12,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 3,
+        "name": "Swedana (Herbal Steam)",
+        "requiredItems": [
+            {
+                "stockId": 14,
+                "quantity": 1
+            },
+            {
+                "stockId": 7,
+                "quantity": 300
+            },
+            {
+                "stockId": 16,
+                "quantity": 2
+            }
+        ]
+    },
+    {
+        "id": 4,
+        "name": "Nasya (Nasal Therapy)",
+        "requiredItems": [
+            {
+                "stockId": 2,
+                "quantity": 20
+            },
+            {
+                "stockId": 25,
+                "quantity": 1
+            },
+            {
+                "stockId": 16,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 5,
+        "name": "Virechana (Purgation)",
+        "requiredItems": [
+            {
+                "stockId": 3,
+                "quantity": 50
+            },
+            {
+                "stockId": 11,
+                "quantity": 20
+            },
+            {
+                "stockId": 16,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 6,
+        "name": "Basti (Enema Therapy)",
+        "requiredItems": [
+            {
+                "stockId": 24,
+                "quantity": 1
+            },
+            {
+                "stockId": 7,
+                "quantity": 200
+            },
+            {
+                "stockId": 16,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 7,
+        "name": "Netra Tarpana (Eye Therapy)",
+        "requiredItems": [
+            {
+                "stockId": 4,
+                "quantity": 50
+            },
+            {
+                "stockId": 26,
+                "quantity": 1
+            },
+            {
+                "stockId": 16,
+                "quantity": 1
+            }
+        ]
+    },
+    {
+        "id": 8,
+        "name": "Pinda Sweda (Rice Bolus Massage)",
+        "requiredItems": [
+            {
+                "stockId": 27,
+                "quantity": 200
+            },
+            {
+                "stockId": 28,
+                "quantity": 100
+            },
+            {
+                "stockId": 29,
+                "quantity": 2
+            },
+            {
+                "stockId": 16,
+                "quantity": 2
+            }
+        ]
+    },
+    {
+        "id": 9,
+        "name": "Udvartana (Powder Massage)",
+        "requiredItems": [
+            {
+                "stockId": 5,
+                "quantity": 100
+            },
+            {
+                "stockId": 12,
+                "quantity": 1
+            },
+            {
+                "stockId": 16,
+                "quantity": 2
+            }
+        ]
+    },
+    {
+        "id": 10,
+        "name": "Karna Purana (Ear Therapy)",
+        "requiredItems": [
+            {
+                "stockId": 1,
+                "quantity": 20
+            },
+            {
+                "stockId": 16,
+                "quantity": 1
+            }
+        ]
+    }
+]
